Module Grpc_quickcheck

Running roundtrip tests for RPC with quickcheck.

module type S = sig ... end

type ('request, 'request_mode, 'response, 'response_mode) t = (module S with type request_mode = 'request_mode and type response_mode = 'response_mode and type Request.t = 'request and type Response.t = 'response)

val run_exn : Base.Source_code_position.t -> ?config:Base_quickcheck.Test.Config.t -> ?requests:'request Base.list -> ?responses:'response Base.list -> ('request, 'request_mode, 'response, 'response_mode) t -> Base.unit

Run 2 quickcheck tests to go over requests and responses and make sure the generated inputs roundtrips correctly through serialization.

Individual tests

val run_request_exn : Base.Source_code_position.t -> ?config:Base_quickcheck.Test.Config.t -> ?examples:'request Base.list -> ('request, 'request_mode, 'response, 'response_mode) t -> Base.unit

val run_response_exn : Base.Source_code_position.t -> ?config:Base_quickcheck.Test.Config.t -> ?examples:'response Base.list -> ('request, 'request_mode, 'response, 'response_mode) t -> Base.unit